Added some very basic tests for config schema testing
